Got rid of SimpleConfig, move LayeredConfig and related classes under core
