+++ src/json_config/api/__init__.py
@@ -1,12 +1,10 @@
-from ..layered_config.config import ConfigValues, LayeredConfig
-from ..layered_config.config_layer import ConfigLayer
-from ..layered_config.config_manager import LayeredConfigManager
-from ..simple_config import SimpleConfig
+from ..core.config import ConfigValues, LayeredConfig
+from ..core.config_layer import ConfigLayer
+from ..core.config_manager import LayeredConfigManager
 
 __all__ = [
-    "LayeredConfig",
     "ConfigLayer",
-    "LayeredConfigManager",
-    "SimpleConfig",
     "ConfigValues",
+    "LayeredConfig",
+    "LayeredConfigManager",
 ]
