+++ tests/core/test_config_values.py
